fix: address PR comments, split into separate utils file, clarifying comments, performant date parsing
