+++ .github/workflows/test.yml
@@ -13,15 +13,3 @@ jobs:
       - run: sudo apt install -y libcurl4-openssl-dev libpcre++-dev
       - run: zig build
       - run: zig build test
-      - name: Set up Python 3.7
-        uses: actions/setup-python@v1
-        with:
-          python-version: 3.7
-      - name: Install pylint and isort
-        run: |
-          python -m pip install --upgrade pip
-          pip install pylint isort
-      - name: Run pylint and isort
-        run: |
-          isort --check-only **/*.py
-          pylint **/*.py
